cubestat: metrics v1 (disk)
